+++ python/orca/src/bigdl/orca/learn/utils.py
@@ -449,113 +449,113 @@ def get_checkpoint_from_callbacks(callbacks):
     return None
 
 
-def put_local_dir_to_remote(local_dir, remote_dir):
-    if remote_dir.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        host_port = remote_dir.split("://")[1].split("/")[0].split(":")
-        classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
-                                     stdout=subprocess.PIPE).communicate()[0]
-        os.environ["CLASSPATH"] = classpath.decode("utf-8")
-        fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
-        if not fs.exists(remote_dir):
-            fs.mkdir(remote_dir)
-        for file in os.listdir(local_dir):
-            with open(os.path.join(local_dir, file), "rb") as f:
-                fs.upload(os.path.join(remote_dir, file), f)
-    elif remote_dir.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = remote_dir.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        prefix = "/".join(path_parts)
-        for file in os.listdir(local_dir):
-            with open(os.path.join(local_dir, file), "rb") as f:
-                s3_client.upload_fileobj(f, Bucket=bucket, Key=prefix+'/'+file)
-    else:
-        if remote_dir.startswith("file://"):
-            remote_dir = remote_dir[len("file://"):]
-        from distutils.dir_util import copy_tree
-        copy_tree(local_dir, remote_dir)
+# def put_local_dir_to_remote(local_dir, remote_dir):
+#     if remote_dir.startswith("hdfs"):  # hdfs://url:port/file_path
+#         import pyarrow as pa
+#         host_port = remote_dir.split("://")[1].split("/")[0].split(":")
+#         classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
+#                                      stdout=subprocess.PIPE).communicate()[0]
+#         os.environ["CLASSPATH"] = classpath.decode("utf-8")
+#         fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
+#         if not fs.exists(remote_dir):
+#             fs.mkdir(remote_dir)
+#         for file in os.listdir(local_dir):
+#             with open(os.path.join(local_dir, file), "rb") as f:
+#                 fs.upload(os.path.join(remote_dir, file), f)
+#     elif remote_dir.startswith("s3"):  # s3://bucket/file_path
+#         access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
+#         secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
+#         import boto3
+#         s3_client = boto3.Session(
+#             aws_access_key_id=access_key_id,
+#             aws_secret_access_key=secret_access_key).client('s3', verify=False)
+#         path_parts = remote_dir.split("://")[1].split('/')
+#         bucket = path_parts.pop(0)
+#         prefix = "/".join(path_parts)
+#         for file in os.listdir(local_dir):
+#             with open(os.path.join(local_dir, file), "rb") as f:
+#                 s3_client.upload_fileobj(f, Bucket=bucket, Key=prefix+'/'+file)
+#     else:
+#         if remote_dir.startswith("file://"):
+#             remote_dir = remote_dir[len("file://"):]
+#         from distutils.dir_util import copy_tree
+#         copy_tree(local_dir, remote_dir)
 
-def get_remote_dir_to_local(remote_dir, local_dir):
-    if remote_dir.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        host_port = remote_dir.split("://")[1].split("/")[0].split(":")
-        classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
-                                     stdout=subprocess.PIPE).communicate()[0]
-        os.environ["CLASSPATH"] = classpath.decode("utf-8")
-        fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
-        if not fs.exists(remote_dir):
-            raise Exception("The remote directory {} does not exist".format(remote_dir))
-        for file in fs.ls(remote_dir):
-            base_name = os.path.basename(file)
-            local_file = os.path.join(local_dir, base_name)
-            with open(local_file, "wb") as f:
-                fs.download(os.path.join(remote_dir, file), f)
-    elif remote_dir.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = remote_dir.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        prefix = "/".join(path_parts)
-        resp = s3_client.list_objects_v2(Bucket=bucket,
-                                         Prefix=prefix)
-        for obj in resp['Contents']:
-            key = obj['Key']
-            file_name = os.path.basename(key)
-            s3_client.download_file(bucket, key, os.path.join(local_dir, file_name))
-    else:
-        if remote_dir.startswith("file://"):
-            remote_dir = remote_dir[len("file://"):]
-        import shutil
-        shutil.copytree(remote_dir, local_dir)
+# def get_remote_dir_to_local(remote_dir, local_dir):
+#     if remote_dir.startswith("hdfs"):  # hdfs://url:port/file_path
+#         import pyarrow as pa
+#         host_port = remote_dir.split("://")[1].split("/")[0].split(":")
+#         classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
+#                                      stdout=subprocess.PIPE).communicate()[0]
+#         os.environ["CLASSPATH"] = classpath.decode("utf-8")
+#         fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
+#         if not fs.exists(remote_dir):
+#             raise Exception("The remote directory {} does not exist".format(remote_dir))
+#         for file in fs.ls(remote_dir):
+#             base_name = os.path.basename(file)
+#             local_file = os.path.join(local_dir, base_name)
+#             with open(local_file, "wb") as f:
+#                 fs.download(os.path.join(remote_dir, file), f)
+#     elif remote_dir.startswith("s3"):  # s3://bucket/file_path
+#         access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
+#         secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
+#         import boto3
+#         s3_client = boto3.Session(
+#             aws_access_key_id=access_key_id,
+#             aws_secret_access_key=secret_access_key).client('s3', verify=False)
+#         path_parts = remote_dir.split("://")[1].split('/')
+#         bucket = path_parts.pop(0)
+#         prefix = "/".join(path_parts)
+#         resp = s3_client.list_objects_v2(Bucket=bucket,
+#                                          Prefix=prefix)
+#         for obj in resp['Contents']:
+#             key = obj['Key']
+#             file_name = os.path.basename(key)
+#             s3_client.download_file(bucket, key, os.path.join(local_dir, file_name))
+#     else:
+#         if remote_dir.startswith("file://"):
+#             remote_dir = remote_dir[len("file://"):]
+#         import shutil
+#         shutil.copytree(remote_dir, local_dir)
 
-def get_remote_files_to_local_dir(remote_path_prefix, local_dir):
-    if remote_path_prefix.startswith("hdfs"):  # hdfs://url:port/file_path
-        import pyarrow as pa
-        remote_dir = os.path.dirname(remote_path_prefix)
-        prefix = os.path.basename(remote_path_prefix)
-        host_port = remote_dir.split("://")[1].split("/")[0].split(":")
-        classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
-                                     stdout=subprocess.PIPE).communicate()[0]
-        os.environ["CLASSPATH"] = classpath.decode("utf-8")
-        fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
-        if not fs.exists(remote_dir):
-            raise Exception("The remote directory {} does not exist".format(remote_dir))
-        for file in fs.ls(remote_dir):
-            base_name = os.path.basename(file)
-            if base_name.startswith(prefix):
-                local_file = os.path.join(local_dir, base_name)
-                with open(local_file, "wb") as f:
-                    fs.download(os.path.join(remote_dir, file), f)
-        return os.path.join(local_dir, prefix)
-    elif remote_path_prefix.startswith("s3"):  # s3://bucket/file_path
-        access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
-        secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
-        import boto3
-        s3_client = boto3.Session(
-            aws_access_key_id=access_key_id,
-            aws_secret_access_key=secret_access_key).client('s3', verify=False)
-        path_parts = remote_path_prefix.split("://")[1].split('/')
-        bucket = path_parts.pop(0)
-        key = "/".join(path_parts)
-        prefix = os.path.basename(key)
-        resp = s3_client.list_objects_v2(Bucket=bucket,
-                                         Prefix=key)
-        for obj in resp['Contents']:
-            file_key = obj['Key']
-            file_name = os.path.basename(file_key)
-            s3_client.download_file(bucket, file_key, os.path.join(local_dir, file_name))
-        return os.path.join(local_dir, prefix)
-    else:
-        if remote_path_prefix.startswith("file://"):
-            remote_path_prefix = remote_path_prefix[len("file://"):]
-        return remote_path_prefix
+# def get_remote_files_to_local_dir(remote_path_prefix, local_dir):
+#     if remote_path_prefix.startswith("hdfs"):  # hdfs://url:port/file_path
+#         import pyarrow as pa
+#         remote_dir = os.path.dirname(remote_path_prefix)
+#         prefix = os.path.basename(remote_path_prefix)
+#         host_port = remote_dir.split("://")[1].split("/")[0].split(":")
+#         classpath = subprocess.Popen(["hadoop", "classpath", "--glob"],
+#                                      stdout=subprocess.PIPE).communicate()[0]
+#         os.environ["CLASSPATH"] = classpath.decode("utf-8")
+#         fs = pa.hdfs.connect(host=host_port[0], port=int(host_port[1]))
+#         if not fs.exists(remote_dir):
+#             raise Exception("The remote directory {} does not exist".format(remote_dir))
+#         for file in fs.ls(remote_dir):
+#             base_name = os.path.basename(file)
+#             if base_name.startswith(prefix):
+#                 local_file = os.path.join(local_dir, base_name)
+#                 with open(local_file, "wb") as f:
+#                     fs.download(os.path.join(remote_dir, file), f)
+#         return os.path.join(local_dir, prefix)
+#     elif remote_path_prefix.startswith("s3"):  # s3://bucket/file_path
+#         access_key_id = os.environ["AWS_ACCESS_KEY_ID"]
+#         secret_access_key = os.environ["AWS_SECRET_ACCESS_KEY"]
+#         import boto3
+#         s3_client = boto3.Session(
+#             aws_access_key_id=access_key_id,
+#             aws_secret_access_key=secret_access_key).client('s3', verify=False)
+#         path_parts = remote_path_prefix.split("://")[1].split('/')
+#         bucket = path_parts.pop(0)
+#         key = "/".join(path_parts)
+#         prefix = os.path.basename(key)
+#         resp = s3_client.list_objects_v2(Bucket=bucket,
+#                                          Prefix=key)
+#         for obj in resp['Contents']:
+#             file_key = obj['Key']
+#             file_name = os.path.basename(file_key)
+#             s3_client.download_file(bucket, file_key, os.path.join(local_dir, file_name))
+#         return os.path.join(local_dir, prefix)
+#     else:
+#         if remote_path_prefix.startswith("file://"):
+#             remote_path_prefix = remote_path_prefix[len("file://"):]
+#         return remote_path_prefix
